# Stream performance improvements