--- a/src/Core/include/OpenLoco/Core/FileStream.h
+++ b/src/Core/include/OpenLoco/Core/FileStream.h
@@ -4,20 +4,20 @@
 #include <cstdint>
 #include <cstdio>
 #include <filesystem>
-#include <fstream>
 
 namespace OpenLoco
 {
     class FileStream final : public Stream
     {
     private:
-        std::fstream _fstream;
+        FILE* _file{};
         StreamMode _mode{};
         size_t _length{};
         size_t _offset{};
 
     public:
         FileStream(const std::filesystem::path& path, StreamMode mode);
+        ~FileStream();
 
         bool open(const std::filesystem::path& path, StreamMode mode);