--- a/src/Core/include/OpenLoco/Core/MemoryStream.h
+++ b/src/Core/include/OpenLoco/Core/MemoryStream.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "Span.hpp"
 #include "Stream.hpp"
 #include <cstdint>
 #include <cstdio>
@@ -26,6 +27,10 @@ namespace OpenLoco
         const std::byte* data() const;
 
         std::byte* data();
+
+        stdx::span<std::byte> getSpan();
+
+        stdx::span<const std::byte> getSpan() const;
 
         size_t getLength() const noexcept override;