--- a/src/Core/include/OpenLoco/Core/Stream.hpp
+++ b/src/Core/include/OpenLoco/Core/Stream.hpp
@@ -50,14 +50,14 @@ namespace OpenLoco
         virtual void write(const void*, size_t) = 0;
 
         template<typename T>
-        void read(T& dst)
+        void readValue(T& dst)
         {
             // TODO: Ensure the type is a primitive.
             read(&dst, sizeof(T));
         }
 
         template<typename T>
-        void write(const T& src)
+        void writeValue(const T& src)
         {
             // TODO: Ensure the type is a primitive.
             write(&src, sizeof(T));