Results so far:
```
master:
    MemoryStreamTest.testWrite100MiB (1152 ms)
PR:
    MemoryStreamTest.testWrite100MiB (628 ms)

master:
    FileStreamTest.testRead100MiB (5600 ms)
PR:
    FileStreamTest.testRead100MiB (1915 ms)

master:
    FileStreamTest.testWrite100MiB (7526 ms)
PR:
    FileStreamTest.testWrite100MiB (1662 ms)
```
Might be not final, can still turn a knob here and there regarding the buffer size.